# AK: Fix all quadratic-time append-loops over ByteBuffer