@awesomekling 2023-01-09
if we go in this direction, why do we need both `ByteBuffer` and `Vector<u8>`?