@BenWiederhake 2023-01-09
It looks like all of the calls to any of the append and try_append methods should use Vector instead, which would remove most but not quite all instances of ByteBuffer from Serenity.