@BenWiederhake 2023-01-09
I'll leave this issue open for further discussion, but I guess we'll go a different route then.