@BenWiederhake 2023-01-31
I don't have the capacity to finish #16957. I suggest that either someone takes over the merging, or we at least merge this, in order to fix the timeouts in OSS Fuzz and slowness in all the places that cause accidentally-quadratic running times.